fix: make macOS input field scrollable
Optimize the long text input experience for macOS users.

The past implementation of the input field could not scroll (as the iMessage's
input box does) after the input box grew to the maximum height, nor could it
display the complete content in this case, which made it annoying to edit this
long text with in the app.

This commit solves this problem by makes the InputFieldsView scrollable after
it's height increasing to a max.